impr: move unit tests to src folder
