+++ ryot/src/sprites/tests/mod.rs
@@ -1,0 +1,1 @@
+mod sprite_layout_test;
