--- ryot/tests/mod.rs
@@ -1,2 +1,0 @@
-mod appearances;
-mod sprites;
