--- ryot/tests/sprites/mod.rs
@@ -1,1 +1,0 @@
-mod sprite_layout;
